+++ packages/carp_survey_package/test/carp_survey_package_test.dart
@@ -50,20 +50,26 @@ void main() {
     print('#1 : $protocol');
     final studyJson = toJsonString(protocol);
 
-    StudyProtocol protocolFromJson =
-        StudyProtocol.fromJson(json.decode(studyJson) as Map<String, dynamic>);
+    CAMSStudyProtocol protocolFromJson = CAMSStudyProtocol.fromJson(
+        json.decode(studyJson) as Map<String, dynamic>);
     expect(toJsonString(protocolFromJson), equals(studyJson));
     print('#2 : $protocolFromJson');
   });
+
   test('JSON File -> StudyProtocol', () async {
     // Read the study protocol from json file
-    String plainJson = File('test/json/study_1.json').readAsStringSync();
+    String plainJson = File('test/json/study_protocol.json').readAsStringSync();
 
     CAMSStudyProtocol protocol = CAMSStudyProtocol.fromJson(
         json.decode(plainJson) as Map<String, dynamic>);
 
     expect(protocol.ownerId, 'AB');
     expect(protocol.masterDevices.first.roleName, Smartphone.DEFAULT_ROLENAME);
+    expect(
+        protocol.tasks.first.measures
+            .firstWhere((measure) => measure is RPTaskMeasure)
+            .type,
+        "dk.cachet.carp.survey");
     print(toJsonString(protocol));
   });
   test('', () {});